Remove Future#set(null) in fn-execution
Use a sentinel value instead.
