+++ sdks/java/build-tools/src/main/resources/beam/findbugs-filter.xml
@@ -39,47 +39,6 @@
           unapproved artifact license.
 	-->
   <Match>
-    <Class name="org.apache.beam.fn.harness.control.BeamFnControlClient$InboundObserver"/>
-    <Method name="onCompleted"/>
-    <Bug pattern="NP_NONNULL_PARAM_VIOLATION"/>
-    <!--
-    Findbugs limitation when using Void typed CompletableFuture. This is a limitation of FindBugs as discussed here:
-    https://github.com/findbugsproject/findbugs/issues/79
-    -->
-  </Match>
-
-  <Match>
-    <Class name="org.apache.beam.sdk.fn.data.BeamFnDataInboundObserver"/>
-    <Method name="receive"/>
-    <Bug pattern="NP_NONNULL_PARAM_VIOLATION"/>
-    <!--
-    Findbugs limitation when using Void typed CompletableFuture. This is a limitation of FindBugs as discussed here:
-    https://github.com/findbugsproject/findbugs/issues/79
-    -->
-  </Match>
-
-  <Match>
-    <Class name="org.apache.beam.fn.harness.logging.BeamFnLoggingClient$LogControlObserver"/>
-    <Method name="onCompleted"/>
-    <Bug pattern="NP_NONNULL_PARAM_VIOLATION"/>
-    <!--
-    Findbugs limitation when using Void typed CompletableFuture. This is a limitation of FindBugs as discussed here:
-    https://github.com/findbugsproject/findbugs/issues/79
-    -->
-  </Match>
-
-  <Match>
-    <Class name="org.apache.beam.fn.harness.data.CompletableFutureInboundDataClient"/>
-    <Method name="complete"/>
-    <Bug pattern="NP_NONNULL_PARAM_VIOLATION"/>
-    <!--
-    Findbugs limitation when using Void typed CompletableFuture. This is a limitation of FindBugs as discussed here:
-    https://github.com/findbugsproject/findbugs/issues/79
-    -->
-  </Match>
-
-
-  <Match>
     <Class name="org.apache.beam.sdk.coders.AvroCoder$SerializableSchemaSupplier"/>
     <Field name="schema"/>
     <Bug pattern="SE_BAD_FIELD"/>
